Moved gem summary to below paper doll on gear tab.
 Changes to be committed:
	modified:   ui/core/components/gear_picker.tsx
	deleted:    ui/core/components/gem_summary_action.ts
	modified:   ui/core/individual_sim_ui.ts
